Project for ACTS 190- Fitting Distributions
10-15-21 update- Austin:
VERY rough draft of much of the content for the first presentation. much more interpretation is needed. I began the content for presentation 2 so we can have a backup in case we can't make the content last 30 mins. Experiencing errors with fitdist() corresponding to the gamma and exponential distributions. I don't think it is related to the numbers, but more likely the data/ data format itself.